Integrated new subscribePrecacheStaticDependencies event.
