+++ src/offline.ts
@@ -2,10 +2,13 @@ import {
   generateOfflineUuid,
   getSynchronizationItems,
   isOfflineUuid,
+  messageOmrsServiceWorker,
+  openmrsFetch,
   QueueItemDescriptor,
   queueSynchronizationItem,
   setupOfflineSync,
   subscribeNetworkRequestFailed,
+  subscribePrecacheStaticDependencies,
 } from '@openmrs/esm-framework';
 
 export interface QueuedEncounterRequest {
@@ -45,6 +48,25 @@ export function setupEncounterRequestInterceptors() {
         await queueEncounterRequest(encounterRequest);
       }
     }
+  });
+}
+
+export function setupOfflineDataSourcePrecaching() {
+  subscribePrecacheStaticDependencies(async () => {
+    const urlsToCache = [
+      '/ws/rest/v1/location?q=&v=custom:(uuid,display)',
+      '/ws/rest/v1/provider?q=&v=custom:(uuid,display,person:(uuid))',
+    ];
+
+    await Promise.all(
+      urlsToCache.map(async (url) => {
+        await messageOmrsServiceWorker({
+          type: 'registerDynamicRoute',
+          pattern: '.+' + url,
+        });
+        await openmrsFetch(url);
+      }),
+    );
   });
 }
